// frontend/src/types/getItems.ts
export interface GetItems {
    item_id: number;
    item_name: string;
    image: string;
    description: string;
    purchasePrice: number;
    salePrice: number;
    stockQuantity: number;
    minimumStock: number;
    category: string;
    storageLocation: string;
    generalInformation: string;
  }
  